+++ src/iface/interface/mod.rs
@@ -520,321 +520,6 @@ impl Interface {
         poll_at.min()
     }
 
-    #[cfg(feature = "proto-rpl")]
-    fn poll_rpl<D>(&mut self, device: &mut D) -> bool
-    where
-        D: Device + ?Sized,
-    {
-        fn transmit<D>(
-            ctx: &mut InterfaceInner,
-            device: &mut D,
-            packet: IpPacket,
-            fragmenter: &mut Fragmenter,
-        ) -> bool
-        where
-            D: Device + ?Sized,
-        {
-            let Some(tx_token) = device.transmit(ctx.now) else {
-                return false;
-            };
-
-            match ctx.dispatch_ip(tx_token, PacketMeta::default(), packet, fragmenter) {
-                Ok(()) => true,
-                Err(e) => {
-                    net_debug!("failed to send packet: {:?}", e);
-                    false
-                }
-            }
-        }
-
-        let Interface {
-            inner: ctx,
-            fragmenter,
-            ..
-        } = self;
-
-        // When we are not the root and we are not part of any DODAG, make sure to transmit
-        // a DODAG Information Solicitation (DIS) message. Only transmit this message when
-        // the DIS timer is expired.
-        if !ctx.rpl.is_root && ctx.rpl.dodag.is_none() && ctx.now >= ctx.rpl.dis_expiration {
-            net_trace!("transmitting RPL DIS");
-            ctx.rpl.dis_expiration = ctx.now + Duration::from_secs(60);
-
-            let dis = RplRepr::DodagInformationSolicitation(RplDis {
-                options: Default::default(),
-            });
-            let icmp_rpl = Icmpv6Repr::Rpl(dis);
-
-            let ipv6_repr = Ipv6Repr {
-                src_addr: ctx.ipv6_addr().unwrap(),
-                dst_addr: Ipv6Address::LINK_LOCAL_ALL_RPL_NODES,
-                next_header: IpProtocol::Icmpv6,
-                payload_len: icmp_rpl.buffer_len(),
-                hop_limit: 64,
-            };
-
-            return transmit(
-                ctx,
-                device,
-                IpPacket::new_ipv6(ipv6_repr, IpPayload::Icmpv6(icmp_rpl)),
-                fragmenter,
-            );
-        }
-
-        let our_addr = ctx.ipv6_addr().unwrap();
-        let Some(dodag) = &mut ctx.rpl.dodag else {
-            return false;
-        };
-
-        // Remove stale relations and increment the DTSN when we actually removed
-        // a relation. This means that a node forgot to retransmit a DAO on time.
-        #[cfg(any(feature = "rpl-mop-1", feature = "rpl-mop-2", feature = "rpl-mop-3"))]
-        if dodag.relations.purge(ctx.now) {
-            use crate::iface::rpl::ModeOfOperation;
-            if match ctx.rpl.mode_of_operation {
-                #[cfg(feature = "rpl-mop-1")]
-                ModeOfOperation::NonStoringMode if ctx.rpl.is_root => true,
-                #[cfg(feature = "rpl-mop-2")]
-                ModeOfOperation::StoringMode => true,
-                #[cfg(feature = "rpl-mop-3")]
-                ModeOfOperation::StoringModeWithMulticast => true,
-                _ => false,
-            } && dodag.dtsn_incremented_at < dodag.dio_timer.next_expiration()
-            {
-                net_trace!("incrementing DTSN");
-                dodag.dtsn_incremented_at = dodag.dio_timer.next_expiration();
-                dodag.dtsn.increment();
-            }
-        }
-
-        // Schedule a DAO before the route will expire.
-        if let Some(parent_address) = dodag.parent {
-            let parent = dodag.parent_set.find(&parent_address).unwrap();
-
-            // If we did not hear from our parent for some time,
-            // remove our parent. Ideally, we should check if we could find another parent.
-            if parent.last_heard < ctx.now - dodag.dio_timer.max_expiration() * 2 {
-                dodag.remove_parent(
-                    ctx.rpl.mode_of_operation,
-                    our_addr,
-                    &ctx.rpl.of,
-                    ctx.now,
-                    &mut ctx.rand,
-                );
-
-                net_trace!("transmitting DIO (INFINITE rank)");
-                let mut options = heapless::Vec::new();
-                options.push(ctx.rpl.dodag_configuration()).unwrap();
-
-                let icmp = Icmpv6Repr::Rpl(ctx.rpl.dodag_information_object(options));
-
-                let ipv6_repr = Ipv6Repr {
-                    src_addr: ctx.ipv6_addr().unwrap(),
-                    dst_addr: Ipv6Address::LINK_LOCAL_ALL_RPL_NODES,
-                    next_header: IpProtocol::Icmpv6,
-                    payload_len: icmp.buffer_len(),
-                    hop_limit: 64,
-                };
-
-                ctx.rpl.dodag = None;
-                ctx.rpl.dis_expiration = ctx.now + Duration::from_secs(5);
-
-                return transmit(
-                    ctx,
-                    device,
-                    IpPacket::new_ipv6(ipv6_repr, IpPayload::Icmpv6(icmp)),
-                    fragmenter,
-                );
-            }
-
-            #[cfg(any(feature = "rpl-mop-1", feature = "rpl-mop-2", feature = "rpl-mop-3"))]
-            if dodag.dao_expiration <= ctx.now {
-                dodag.schedule_dao(ctx.rpl.mode_of_operation, our_addr, parent_address, ctx.now);
-            }
-        }
-
-        // Transmit any DAO-ACK that are queued.
-        #[cfg(any(feature = "rpl-mop-1", feature = "rpl-mop-2", feature = "rpl-mop-3"))]
-        if !dodag.dao_acks.is_empty() {
-            // Transmit all the DAO-ACKs that are still queued.
-            net_trace!("transmit DAO-ACK");
-
-            #[allow(unused_mut)]
-            let (mut dst_addr, sequence) = dodag.dao_acks.pop().unwrap();
-            let rpl_instance_id = dodag.instance_id;
-            let icmp = Icmpv6Repr::Rpl(RplRepr::DestinationAdvertisementObjectAck(RplDaoAck {
-                rpl_instance_id,
-                sequence,
-                status: 0,
-                dodag_id: if rpl_instance_id.is_local() {
-                    Some(dodag.id)
-                } else {
-                    None
-                },
-            }));
-
-            let mut options = heapless::Vec::new();
-            options
-                .push(Ipv6OptionRepr::Rpl(RplHopByHopRepr {
-                    down: true,
-                    rank_error: false,
-                    forwarding_error: false,
-                    instance_id: ctx.rpl.dodag.as_ref().unwrap().instance_id,
-                    sender_rank: ctx.rpl.dodag.as_ref().unwrap().rank.raw_value(),
-                }))
-                .unwrap();
-            #[allow(unused_mut)]
-            let mut hop_by_hop = Some(Ipv6HopByHopRepr { options });
-
-            // A DAO-ACK always goes down. In MOP1, both Hop-by-Hop option and source
-            // routing header MAY be included. However, a source routing header must always
-            // be included when it is going down.
-            #[cfg(feature = "rpl-mop-1")]
-            let routing = if matches!(
-                ctx.rpl.mode_of_operation,
-                super::RplModeOfOperation::NonStoringMode
-            ) && ctx.rpl.is_root
-            {
-                net_trace!("creating source routing header to {}", dst_addr);
-                if let Some((source_route, new_dst_addr)) =
-                    self::rpl::create_source_routing_header(ctx, our_addr, dst_addr)
-                {
-                    hop_by_hop = None;
-                    dst_addr = new_dst_addr;
-                    Some(source_route)
-                } else {
-                    None
-                }
-            } else {
-                None
-            };
-
-            #[cfg(not(feature = "rpl-mop-1"))]
-            let routing = None;
-
-            let ip_packet = super::ip_packet::Ipv6Packet {
-                header: Ipv6Repr {
-                    src_addr: ctx.ipv6_addr().unwrap(),
-                    dst_addr,
-                    next_header: IpProtocol::Icmpv6,
-                    payload_len: icmp.buffer_len(),
-                    hop_limit: 64,
-                },
-                hop_by_hop,
-                routing,
-                payload: IpPayload::Icmpv6(icmp),
-            };
-            return transmit(ctx, device, IpPacket::Ipv6(ip_packet), fragmenter);
-        }
-
-        // Transmit any DAO that are queued.
-        #[cfg(any(feature = "rpl-mop-1", feature = "rpl-mop-2", feature = "rpl-mop-3"))]
-        if !dodag.daos.is_empty() {
-            // Remove DAOs that have been transmitted 3 times and did not get acknowledged.
-            // TODO: we should be able to remove the parent when it was actually not acknowledged
-            // after 3 times. This means that there is no valid path to the parent.
-            dodag.daos.retain(|dao| {
-                (!dao.is_no_path && dao.sent_count < 4) || (dao.is_no_path && dao.sent_count == 0)
-            });
-
-            // Go over each queued DAO and check if they need to be transmitted.
-            dodag.daos.iter_mut().for_each(|dao| {
-                if !dao.needs_sending {
-                    let Some(next_tx) = dao.next_tx else {
-                        dao.next_tx = Some(ctx.now + dodag.dio_timer.min_expiration());
-                        return;
-                    };
-
-                    if next_tx < ctx.now {
-                        dao.needs_sending = true;
-                    }
-                }
-            });
-
-            if let Some(dao) = dodag.daos.iter_mut().find(|dao| dao.needs_sending) {
-                dao.next_tx = Some(ctx.now + Duration::from_secs(60));
-                dao.sent_count += 1;
-                dao.needs_sending = false;
-                let dst_addr = dao.to;
-
-                let icmp = Icmpv6Repr::Rpl(dao.as_rpl_dao_repr());
-
-                let mut options = heapless::Vec::new();
-                options
-                    .push(Ipv6OptionRepr::Rpl(RplHopByHopRepr {
-                        down: ctx.rpl.is_root,
-                        rank_error: false,
-                        forwarding_error: false,
-                        instance_id: dodag.instance_id,
-                        sender_rank: dao.rank.raw_value(),
-                    }))
-                    .unwrap();
-
-                let hbh = Ipv6HopByHopRepr { options };
-
-                let ip_packet = super::ip_packet::Ipv6Packet {
-                    header: Ipv6Repr {
-                        src_addr: our_addr,
-                        dst_addr,
-                        next_header: IpProtocol::Icmpv6,
-                        payload_len: icmp.buffer_len(),
-                        hop_limit: 64,
-                    },
-                    hop_by_hop: Some(hbh),
-                    routing: None,
-                    payload: IpPayload::Icmpv6(icmp),
-                };
-                net_trace!("transmitting DAO");
-                return transmit(ctx, device, IpPacket::Ipv6(ip_packet), fragmenter);
-            }
-        }
-
-        // When we are part of a DODAG, we should check if our DIO Trickle timer
-        // expired. If it expires, a DODAG Information Object (DIO) should be
-        // transmitted.
-        if (ctx.rpl.is_root || dodag.parent.is_some())
-            && dodag.dio_timer.poll(ctx.now, &mut ctx.rand)
-        {
-            let mut options = heapless::Vec::new();
-            options.push(ctx.rpl.dodag_configuration()).unwrap();
-
-            let icmp = Icmpv6Repr::Rpl(ctx.rpl.dodag_information_object(options));
-
-            let ipv6_repr = Ipv6Repr {
-                src_addr: our_addr,
-                dst_addr: Ipv6Address::LINK_LOCAL_ALL_RPL_NODES,
-                next_header: IpProtocol::Icmpv6,
-                payload_len: icmp.buffer_len(),
-                hop_limit: 64,
-            };
-
-            return transmit(
-                ctx,
-                device,
-                IpPacket::new_ipv6(ipv6_repr, IpPayload::Icmpv6(icmp)),
-                fragmenter,
-            );
-        }
-
-        false
-    }
-
-    #[cfg(feature = "proto-rpl")]
-    fn poll_at_rpl(&mut self) -> Instant {
-        let ctx = self.context();
-
-        if let Some(dodag) = &ctx.rpl.dodag {
-            #[cfg(any(feature = "rpl-mop-1", feature = "rpl-mop-2", feature = "rpl-mop-3"))]
-            if !dodag.daos.is_empty() || !dodag.dao_acks.is_empty() {
-                return Instant::from_millis(0);
-            }
-
-            dodag.dio_timer.poll_at()
-        } else {
-            ctx.rpl.dis_expiration
-        }
-    }
-
     /// Return an _advisory wait time_ for calling [poll] the next time.
     /// The [Duration] returned is the time left to wait before calling [poll] next.
     /// It is harmless (but wastes energy) to call it before the [Duration] has passed,
